refactor:フィルタ実行部分を削除
move:UnsharpMaskingフォルダに移動
